+++ contracts/interfaces/ISynchronizer.sol
@@ -25,7 +25,7 @@ interface ISynchronizer {
     );
     event WithdrawFee(address partner, uint256 partnerFee, uint256 platformFee, uint256 registrarType);
     event SetMinimumRequiredSignatures(uint256 oldValue, uint256 newValue);
-    event SetAppId(uint8 oldId, uint8 newId);
+    event SetAppId(uint32 oldId, uint32 newId);
     event SetVirtualReserve(uint256 oldReserve, uint256 newReserve);
     event SetMuonContract(address oldContract, address newContract);
     event ToggleUseVirtualReserve(bool useVirtualReserve);
